factor out agentOpts, main -> commonMain, 2x GOOS=windows guards for root/devmode, factor out WaitOnline, factor out SIGURG
